+++ src/cli/__init__.py
@@ -1,8 +1,3 @@
 """命令行界面模块"""
 
-try:
-    from .main import cli
-except ImportError:
-    from cli.main import cli
-
 __all__ = ['cli']